requirements.txt
Jinja2==3.1.6
python-frontmatter==1.1.0
humanize==4.12.3
markdown==3.8.2
pygments==2.19.2